Move core_number to its own file
